+++ config/livewire.php
@@ -3,13 +3,17 @@
 return [
     'temporary_file_upload' => [
         'disk' => env('LIVEWIRE_TEMPORARY_FILE_UPLOAD_DISK', 'local'),
-        'rules' => null,
+        'rules' => [
+            'required',
+            'file',
+            'mimes:jpg,jpeg,png,webp',
+            'max:2048',
+            'dimensions:max_width=5000,max_height=5000',
+        ],
         'directory' => null,
-        'middleware' => null,
+        'middleware' => 'throttle:20,1',
         'preview_mimes' => [
-            'png', 'gif', 'bmp', 'svg', 'wav', 'mp4',
-            'mov', 'avi', 'wmv', 'mp3', 'm4a',
-            'jpg', 'jpeg', 'mpga', 'webp', 'wma',
+            'png', 'jpg', 'jpeg', 'webp',
         ],
         'max_upload_time' => 5,
         'cleanup' => true,